refactor: extract extensions AppModel & add CustomStringConvertible
